requirements.txt
arrow~=1.2.2
Jinja2~=3.0.3
PyYAML~=6.0
